fix sanity scripts

+++ src/emp/datacentre/sanity_calculate_util.sh
@@ -13,9 +13,15 @@ for seed in $(seq $seedfrom $seedto)
 do
   outputfile=${outputfileprefix}_${seed}
   fctoutputfile=${outputfile}_fct
-  cat $outputfile | grep -e "FCT" > $fctoutputfile
-  python pre_aux_parse_ls_report_generically.py $outputfile
   parsedoutputfile=${outputfileprefix}_${seed}_parsed
+  if [ ! -s $fctoutputfile ]
+  then
+    cat $outputfile | grep -e "FCT" > $fctoutputfile
+  fi
+  if [ ! -s $parsedoutputfile ]
+  then
+    python pre_aux_parse_ls_report_generically.py $outputfile
+  fi
   python sanity_aux_calculate_network_util.py $parsedoutputfile $mstart $mend > $tempoutputfile
   utillower=$(cat $tempoutputfile | cut -d " " -f 1)
   utilupper=$(cat $tempoutputfile | cut -d " " -f 2)
